Start on a keyframe 'core' for loading a frame, and also add an 'error' vector element type

+++ animation/src/storage/storage_api.rs
@@ -100,5 +100,5 @@ pub enum StorageResponse {
     KeyFrame(Duration, Duration),
 
     /// The serialized version of the element that was requested
-    Element(usize, String)
+    Element(i64, String)
 }
